Fix shift-overflow and logical-op warnings
- Fix shift-overflow in MouseEvents.c by casting to UInt32 before shift
- Fix redundant logical-or in folder_window.c extension check

+++ src/EventManager/MouseEvents.c
@@ -841,7 +841,7 @@ EventRecord GenerateMouseMovedEvent(Point position, SInt16 modifiers)
     EventRecord event = {0};
 
     event.what = osEvt;
-    event.message = mouseMovedMessage << 24;
+    event.message = (UInt32)mouseMovedMessage << 24;
     event.when = TickCount();
     event.where = position;
     event.modifiers = modifiers;
